Refactor to gradle project
