+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Wed Jun 25 08:42:44 BRT 2025
+gradle.version=8.10
