--- src/com/github/theprogmatheus/minecraft/utils/guibuilder/GuiLayout.java
@@ -1,10 +1,0 @@
-package com.github.theprogmatheus.minecraft.utils.guibuilder;
-
-public interface GuiLayout
-{
-
-	public abstract int[] slots();
-
-	public abstract int size();
-
-}
