--- src/com/github/theprogmatheus/minecraft/utils/guibuilder/GuiRunnable.java
@@ -1,9 +1,0 @@
-package com.github.theprogmatheus.minecraft.utils.guibuilder;
-
-import org.bukkit.entity.Player;
-
-public interface GuiRunnable
-{
-
-	public abstract void run(Player player);
-}
