+++ src/main/java/com/github/theprogmatheus/minecraft/utils/guibuilder/GuiRunnable.java
@@ -1,0 +1,9 @@
+package com.github.theprogmatheus.minecraft.utils.guibuilder;
+
+import org.bukkit.entity.Player;
+
+public interface GuiRunnable
+{
+
+	public abstract void run(Player player);
+}
